Move HTTPServer back into the try block
This is a bit esoteric but it should allow us to correctly re-bind the port when there is a problem binding it. There is still an edge case where if the port bind files and the user sends Ctrl-C the program will hit an exception because httpd=None does not have the shutdown() method but this should either be a vanishingly small edge case or be impossible.